Add cross-platform service management and home directory consistency
- Implement cross-platform service management with ServiceManager for Linux systemd, Windows Service Manager, and macOS launchd
- Add PlatformPaths utility for consistent directory structure across platforms using ~/.aetheric
- Create service_simple.rs module providing unified service management API without external dependencies
- Add MqttBrokerManager for platform-specific Mosquitto installation (apt/chocolatey/homebrew)
- Create install.ps1 PowerShell script for Windows installation with proper service configuration
- Update setup command to use home directories instead of system directories, removing root requirement for configuration
- Fix various clippy linting issues including needless_borrows_for_generic_args and redundant_pattern_matching
- Maintain backward compatibility while providing consistent user experience across all platforms

+++ src/agent/ssh_tunnel.rs
@@ -197,7 +197,7 @@ impl SshTunnelManager {
         // Check session limits
         {
             let sessions = self.sessions.read().await;
-            if sessions.len() >= self.config.ssh.max_sessions as usize {
+            if sessions.len() >= self.config.ssh.max_sessions {
                 return Ok(SshResponse {
                     session_id: session_id.clone(),
                     status: SshSessionStatus::Failed,
